+++ packages/nano_domain/lib/src/nano_domain.dart
@@ -10,6 +10,7 @@ export 'learning/learning_insights.dart';
 export 'learning/refresh_checkpoint.dart';
 export 'quiz/question_bank.dart';
 export 'quiz/junior_quiz_flow.dart';
+export 'quiz/quiz_attempt.dart';
 export 'quiz/senior_quiz_flow.dart';
 export 'quiz/topic_quiz.dart';
 export 'learning/topic_action.dart';